cs336-final-be/requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pinecone-client==2.2.4
pinecone-plugin-inference==2.0.0
python-dotenv==1.0.0
typing-extensions==4.9.0